+++ packages/angular/table/src/index.ts
@@ -1,2 +1,2 @@
-export { Table } from '../table'
-export { TableHeaderCell } from '../table-header-cell'
+export { Table } from './table'
+export { TableHeaderCell } from './table-header-cell'
